removed highlevel and created a file of startup state
